Make Go to Top button a FAB
Fixes #111
