--- src/components/NavigationToolbars/Items/Top.vue
@@ -1,32 +1,0 @@
-<template>
-  <v-tooltip bottom>
-    <template v-slot:activator="{ on: toolTip }">
-      <v-btn
-        v-if="isSortingPage"
-        text
-        icon
-        class="d-none d-md-flex"
-        v-on="{ ...toolTip }"
-        @click="jumpToTop"
-      >
-        <v-icon>mdi-arrow-up</v-icon>
-      </v-btn>
-    </template>
-    <span>{{ $t('actions.toPageTop') }}</span>
-  </v-tooltip>
-</template>
-
-<script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-
-@Component
-export default class Top extends Vue {
-  private get isSortingPage(): boolean {
-    return this.$route.meta && this.$route.meta.sortingPage;
-  }
-
-  private jumpToTop(): void {
-    window.scrollTo(0, 0);
-  }
-}
-</script>
